requirements.txt
requests==2.32.5
bs4==0.0.2
telebot==0.0.5
pillow==11.3.0
cairosvg==2.8.2
dotenv==0.9.9